+++ opendc-web/opendc-web-ui/.gitignore
@@ -24,7 +24,4 @@ yarn-error.log*
 # Environment variables
 .env.local
 
-# Sass output
-*.css
-
 /.next
