+++ opendc-web/opendc-web-ui/src/components/projects/FilterPanel.module.css
@@ -1,0 +1,7 @@
+.filterPanel {
+    display: flex;
+}
+
+.filterPanel > button {
+    flex: 1 !important;
+}
