--- opendc-web/opendc-web-ui/src/components/projects/FilterPanel.module.scss
@@ -1,7 +1,0 @@
-.filterPanel {
-    display: flex;
-
-    button {
-        flex: 1 !important;
-    }
-}
