+++ opendc-web/opendc-web-ui/src/components/topologies/map/controls/ScaleIndicator.module.css
